==> Budget/Py_budget/2.) WF_Budget_2025.ipynb <==
from datetime import datetime as dt  #to manipulate dates

import numpy as np  #to cleanse data
import openpyxl
import pandas as pd
import plotly.express as px  #to create interactive charts
import plotly.graph_objects as go  #to create interactive charts
from datetime import datetime

ytd_data = pd.read_csv(
    'D:\\Sean\\Documents\\Personal\\Budget\\YTD_downloads\\2025\\wellsfargo.csv', header=None, names=['Date','Amount', 'delete1', 'delete2', 'Description'], parse_dates=['Date']
)

del ytd_data['delete1']
del ytd_data['delete2']

ytd_data['Account'] = 'wells_fargo_checking'
ytd_data.rename(columns={'Date':'Trans. Date'}, inplace=True)
ytd_data['Date'] = pd.to_datetime(ytd_data['Trans. Date'].dt.strftime('%m/%Y'), format='%m/%Y')


ytd_data.tail(10)

# Discover Transactions
mapping = {
    # Direct Deposit
    'income': ['PAYROLL'],
    'family': ['JAYELIN', 'JUSTINA', 'SHARON', 'DORRY'],
    'purchases': ['THERAPY STORES', 'AMAZON', 'NOISEP'],
    'other_people': ['NIKA'],
    'venmo': ['VENMO', 'JONATHAN'],
    'refunds': ['PREARRANGE', 'STUBHUB CONS PAYMENTS 240117 000000017167884 9016004299-SEAN LEWIS'],
    'cell_phone': ['T-Mobile', 'TMOBILE'],
    'la_fitness': ['L A FITNESS'],
    'transfer': ['TRANSFER','MOBILE DEPOSIT : REF NUMBER :314230152812'],
    'student_loan': ['DEPT EDUCATION'],
    'dentist': ['OANH LE'],
    'care_credit': ['Synchrony Bank CC'],
    'fees': ['FEE'],
    'gas': ['SHELL', 'CHEVRON', 'ARCO', '76', 'A&A','PURCHASE AUTHORIZED ON 12/21 PALO ALTO GAS & PALO ALTO CA P000000885077547 CARD 1081','BELMONT ALLIANCE'],
    'laundry': ['Cash eWithdrawal in Branch'],
    'parking': ['REDWOOD CITY PARKI REDWOOD CITY', 'STADIUM GARAGE','PURCHASE AUTHORIZED ON 10/25 STADIUM GARAGE BERKELEY CA S584300071995368 CARD 1081'],
    'ATM': ['ATM'],
    'Vincent': ['Kai Feng'],
    'GT_China': ['Candice Mclemore', 'Alipay', 'CHN', 'Macau', 'HKG'],
    'Australia': ['SYDNEY AUS', 'AUS'],
    'Mexico': ['CIUDAD', 'MEX', 'GetYourGuide'],
    'learning': ['BRITISH SWIM SCHOO'],
    'parking': [r'LAZ P.*', r'Park.*'],
    'Don': ['LEWIS JR', 'DON'],
    'Shevon':['Shevon'],
    'haircut': ['DICE'],
    'Cris' : ['Chris'],
    'general_festival': ['Stockton 9'],
    'misc': ['Lewis Sean'],
    'restaurants': ['West Town Bank', 'PIZZA HUT'],
    'movies': ['AMC'],
    'groceries': ['WEIS', 'BJS'],
    'monitor': ['Best Buy'],
    'sport_stuff': ['BAUER*'],
    'India': ['PURCHASE AUTHORIZED ON 07/29 MinistryofExternal Mumbai IND'],
    'movies': ['CINEPOLIS'],
    'groceries': ['TRADER JOE'],
    'parking': ['METER MTA', 'PAYBYPHONE','CITY OF BURLINGAME'],
    'gym': ['24 HOUR FITNESS']

}
for k, v in mapping.items():
    ytd_data.loc[ytd_data.Description.str.contains('|'.join(v), case=False), 'Short_Desc'] = k


ytd_data[ytd_data['Short_Desc'].isnull()].sort_values(by='Amount', ascending=False)

ytd_data.loc[ytd_data['Short_Desc'].isnull(), 'Short_Desc'] = 'unsure'

reocurring_list = ['cell_phone', 'la_fitness', 'student_loan','care_credit']
ytd_data.loc[ytd_data.Short_Desc.isin(reocurring_list), 'Reoccurring_Flag'] = 'Yes'
ytd_data.loc[~ytd_data.Short_Desc.isin(reocurring_list), 'Reoccurring_Flag'] = 'No'

mapping = {
    'Housing': [],
    'Transportation': ['parking'],
    'Food': ['groceries'],
    'Insurance': [],
    'Utilities': ['cell_phone','gas','laundry'],
    'Medical': ['medical','care_credit','dentist'],
    'Government': ['student_loan'],
    'Savings': [],
    'Personal_Spending': ['la_fitness','ATM','restaurants','learning', 'purchases','haircut','gym','monitor'],
    'Recreation_Entertainment': ['general_festival', 'movies','sport_stuff'],
    'Travel': ['GT_China','Australia','Mexico','air_travel', 'airport','India'],
    'Misc': ['unsure','misc'],
    'People': ['family','venmo','other_people','transfer','Don','Vincent','Shevon','Cris'],
    'Payment_and_Interest': ['fees'],
    
    'Income':['income','refunds']
}

for k, v in mapping.items():
    ytd_data.loc[ytd_data.Short_Desc.isin(v), 'Category_2'] = k

ytd_data[ytd_data['Category_2'].isnull()].sort_values(by='Description', ascending=False).sort_values(by='Short_Desc',
                                                                                                     ascending=False)

x = ytd_data.groupby(['Reoccurring_Flag']).size().reset_index(name='count')

x.sort_values(by='count', ascending=False)

ytd_data.reset_index(inplace=True)
ytd_sorted = ytd_data.sort_values(by=['Trans. Date', 'index'])

# Define the new row with datetime objects for date columns
new_row = pd.DataFrame({
    'index': [0],  # Ensure all entries are in list format to align with DataFrame structure
    'Trans. Date': [datetime.strptime('1/1/2023', '%m/%d/%Y')],  # Matching format and conversion
    'Amount': [616.53],
    'Description': ['Balance'],
    'Account': ['wells_fargo_checking'],
    'Date': [pd.to_datetime('1/1/2023')],  # This is a string representation of the date, so it's okay
    'Short_Desc': ['Balance'],
    'Reoccurring_Flag': ['No'],
    'Category_2': ['Balance'],

}, index=[0])

# Concatenate the new_row DataFrame with the existing ytd_sorted DataFrame
ytd_sorted = pd.concat([new_row, ytd_sorted], ignore_index=True)

# Ensure 'Sort_Date' is also in datetime format for the entire DataFrame
ytd_sorted['Trans. Date'] = pd.to_datetime(ytd_sorted['Trans. Date'])
ytd_sorted

ytd_sorted['CumAmount'] = ytd_sorted['Amount'].cumsum()
ytd_sorted['Year'] = ytd_sorted['Trans. Date'].dt.year

ytd_sorted

ytd_sorted['Counter'] = ytd_sorted.groupby('Short_Desc').cumcount() + 1
ytd_sorted

columns = ['Trans. Date', 'Description', 'Amount', 'Category', 'Short_Desc', 'Reoccurring_Flag',
                     'Category_2', 'Date', 'Year', 'Sort_Date', 'CumAmount', 'Counter', 'Account']
ytd_sorted['Category'] = 'NA'
ytd_sorted['Sort_Date'] = ytd_sorted['Trans. Date']

wf = ytd_sorted[columns]

wf.to_excel('D:\\Sean\\Documents\\Personal\\Budget\\2025_output\\WF_Curated_Bills.xlsx')
