==> Budget/Py_budget/5.) Combine.ipynb <==
from datetime import datetime as dt  #to manipulate dates

import numpy as np  #to cleanse data
import openpyxl
import pandas as pd
import plotly.express as px  #to create interactive charts
import plotly.graph_objects as go  #to create interactive charts
from datetime import datetime


wf_data = pd.read_excel(
    'D:\\Sean\\Documents\\Personal\\Budget\\2025_output\\WF_Curated_Bills.xlsx'
)

discover_data = pd.read_excel(
    'D:\\Sean\\Documents\\Personal\\Budget\\2025_output\\Curated_Bills.xlsx'
)

sofi_savings_data = pd.read_excel(
    'D:\\Sean\\Documents\\Personal\\Budget\\2025_output\\sofi_savings_Curated_Bills.xlsx'
)
sofi_checking_data = pd.read_excel(
    'D:\\Sean\\Documents\\Personal\\Budget\\2025_output\\sofi_checking_Curated_Bills.xlsx'
)


combined = pd.concat([wf_data, discover_data, sofi_savings_data,sofi_checking_data], axis=0)
combined.columns = ['ORIG_INDEX','Trans_Date','Description','Amount','Category','Short_Desc','Reoccurring_Flag','Category_2','Date','Year','Sort_Date','CumAmount','Counter','Account']
combined.reset_index(drop=True, inplace=True)
# Assuming 'combined' is your DataFrame
# Convert the datetime columns to date format (without time)
combined['Trans_Date'] = pd.to_datetime(combined['Trans_Date']).dt.date
combined['Date'] = pd.to_datetime(combined['Date']).dt.date
combined['Sort_Date'] = pd.to_datetime(combined['Sort_Date']).dt.date

# Now write to Excel
combined.to_excel('D:\\Sean\\Documents\\Personal\\Budget\\2025_output\\All_Bills.xlsx', index=True)

display(combined)

import pandas as pd

# Assuming 'combined' is your DataFrame
writer = pd.ExcelWriter('D:\\Sean\\Documents\\Personal\\Budget\\2025_output\\All_Bills.xlsx', engine='xlsxwriter')

# Write the DataFrame to the Excel file
combined.to_excel(writer, index=True)

# Access the workbook and worksheet
workbook = writer.book
worksheet = writer.sheets['Sheet1']  # Change 'Sheet1' if you have a different sheet name

# Get the column indices for the date columns (replace with your actual date column names)
date_format = workbook.add_format({'num_format': 'yyyy-mm-dd'})  # Customize the date format as needed

# Apply the date format to the relevant columns
# Example for applying the format to column B (assuming it's a date column), adjust as needed
worksheet.set_column('C:C', None, date_format)  # Adjust 'B:B' based on the actual column
worksheet.set_column('J:J', None, date_format)  # Adjust 'B:B' based on the actual column
worksheet.set_column('L:L', None, date_format)  # Adjust 'B:B' based on the actual column
worksheet.set_column('P:P', None, date_format)  # Adjust 'B:B' based on the actual column

writer.close()
